fix: image loader for the static export
since we're going to serve the website statically, `next image optimization` isn't available for us.
to solve that we can either use a custom loader and let go of the image optimization or we can use a third-party package and optimize the images on build time.
For now, I've went with the custom loader solution, and we can discuss the implementation of other solution when needed.